COMP: removing compiler warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=12581
git-svn-id: http://svn.slicer.org/Slicer4/trunk@12581 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/ProstateNav/TransPerinealProstateRobot/vtkMRMLTransPerinealProstateRobotNode.cxx
@@ -315,7 +315,7 @@ void vtkMRMLTransPerinealProstateRobotNode::Copy(vtkMRMLNode *anode)
   }
 
   this->StatusDescriptors.clear();
-  for (int i=0; i<node->StatusDescriptors.size(); i++)
+  for (unsigned int i=0; i<node->StatusDescriptors.size(); i++)
   {    
     this->StatusDescriptors.push_back(node->StatusDescriptors[i]);
   }
@@ -393,7 +393,7 @@ void vtkMRMLTransPerinealProstateRobotNode::ProcessMRMLEvents( vtkObject *caller
 {
   Superclass::ProcessMRMLEvents(caller, event, callData);
 
-  if (this->StatusDescriptors.size()<=STATUS_SCANNER || this->StatusDescriptors.size()<=STATUS_ROBOT)
+  if (this->StatusDescriptors.size()<= (int)STATUS_SCANNER || this->StatusDescriptors.size()<= (int)STATUS_ROBOT)
   {
     vtkErrorMacro("StatusDescriptors does not contain scanner and robot status");
   }
